Split core tests into multiple files and moved default test case
