+++ danceschool/core/tests/defaults.py
