+++ danceschool/discounts/tests.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from datetime import timedelta
 
 from danceschool.core.constants import REG_VALIDATION_STR, updateConstant
-from danceschool.core.utils.tests import DefaultSchoolTestCase
+from danceschool.core.tests.defaults import DefaultSchoolTestCase
 from danceschool.core.models import Invoice, Registration
 
 from .models import (
