+++ danceschool/payments/payatdoor/tests.py
@@ -10,7 +10,7 @@ from cms.models import StaticPlaceholder
 
 from danceschool.core.models import Registration, Invoice
 from danceschool.core.constants import REG_VALIDATION_STR
-from danceschool.core.utils.tests import DefaultSchoolTestCase
+from danceschool.core.tests.defaults import DefaultSchoolTestCase
 
 # from .cms_plugins import PayAtDoorFormPlugin
 from .constants import ATTHEDOOR_PAYMENTMETHOD_CHOICES
